# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API server for managing Conduit exchanges",
        "title": "Conduit API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/tonyd33/conduit",
            "email": "support@conduit.example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/api/v1",
    "paths": {
        "/exchanges": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all exchanges, optionally filtered by namespace",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "List exchanges",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace to filter exchanges",
                        "name": "namespace",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ListExchangesResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new exchange with the specified configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Create an exchange",
                "parameters": [
                    {
                        "description": "Exchange configuration",
                        "name": "exchange",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.CreateExchangeRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ExchangeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/exchanges/{namespace}/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a specific exchange by namespace and name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Get an exchange",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exchange name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ExchangeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a specific exchange by namespace and name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "exchanges"
                ],
                "summary": "Delete an exchange",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Namespace",
                        "name": "namespace",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exchange name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_apiserver.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/healthz": {
            "get": {
                "description": "Check if the API server is healthy",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "internal_apiserver.ConnectionInfo": {
            "type": "object",
            "properties": {
                "inputSubject": {
                    "type": "string"
                },
                "natsURL": {
                    "description": "NATS connection details",
                    "type": "string"
                },
                "outputSubject": {
                    "type": "string"
                },
                "podName": {
                    "description": "Worker pod information",
                    "type": "string"
                },
                "podPhase": {
                    "type": "string"
                },
                "ready": {
                    "description": "Status",
                    "type": "boolean"
                },
                "streamName": {
                    "description": "Stream and subject information",
                    "type": "string"
                }
            }
        },
        "internal_apiserver.ConsumerConfig": {
            "type": "object",
            "properties": {
                "ackPolicy": {
                    "description": "\"explicit\", \"all\", or \"none\"",
                    "type": "string"
                },
                "ackWait": {
                    "description": "Duration string like \"30s\"",
                    "type": "string"
                },
                "filterSubject": {
                    "type": "string"
                },
                "headersOnly": {
                    "type": "boolean"
                },
                "maxAckPending": {
                    "type": "integer"
                },
                "maxDeliver": {
                    "type": "integer"
                },
                "maxWaiting": {
                    "type": "integer"
                },
                "replayPolicy": {
                    "description": "\"instant\" or \"original\"",
                    "type": "string"
                }
            }
        },
        "internal_apiserver.CreateExchangeRequest": {
            "type": "object",
            "properties": {
                "consumerConfig": {
                    "description": "JetStream consumer configuration",
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_apiserver.ConsumerConfig"
                        }
                    ]
                },
                "env": {
                    "description": "Environment variables for the worker pod",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_apiserver.EnvVar"
                    }
                },
                "image": {
                    "description": "Image for the worker pod (defaults to conduit default image)",
                    "type": "string"
                },
                "imagePullPolicy": {
                    "description": "ImagePullPolicy for the worker pod",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the Exchange (must be unique within namespace)",
                    "type": "string"
                },
                "namespace": {
                    "description": "Namespace to create the Exchange in (defaults to \"default\")",
                    "type": "string"
                },
                "recovery": {
                    "description": "Recovery configuration",
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_apiserver.RecoveryConfig"
                        }
                    ]
                },
                "resources": {
                    "description": "Resource requests and limits",
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_apiserver.ResourceRequirements"
                        }
                    ]
                },
                "streamConfig": {
                    "description": "JetStream stream configuration",
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_apiserver.StreamConfig"
                        }
                    ]
                }
            }
        },
        "internal_apiserver.EnvVar": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "internal_apiserver.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_apiserver.ExchangeResponse": {
            "type": "object",
            "properties": {
                "connectionInfo": {
                    "description": "Connection information for clients",
                    "allOf": [
                        {
                            "$ref": "#/definitions/internal_apiserver.ConnectionInfo"
                        }
                    ]
                },
                "message": {
                    "type": "string"
                },
                "name": {
                    "description": "Exchange metadata",
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "phase": {
                    "description": "Current status",
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "internal_apiserver.ListExchangesResponse": {
            "type": "object",
            "properties": {
                "exchanges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_apiserver.ExchangeResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "internal_apiserver.RecoveryConfig": {
            "type": "object",
            "properties": {
                "backoffLimit": {
                    "type": "integer"
                },
                "maxRestarts": {
                    "type": "integer"
                },
                "restartPolicy": {
                    "description": "\"Always\", \"OnFailure\", or \"Never\"",
                    "type": "string"
                }
            }
        },
        "internal_apiserver.ResourceList": {
            "type": "object",
            "properties": {
                "cpu": {
                    "type": "string"
                },
                "memory": {
                    "type": "string"
                }
            }
        },
        "internal_apiserver.ResourceRequirements": {
            "type": "object",
            "properties": {
                "limits": {
                    "$ref": "#/definitions/internal_apiserver.ResourceList"
                },
                "requests": {
                    "$ref": "#/definitions/internal_apiserver.ResourceList"
                }
            }
        },
        "internal_apiserver.StreamConfig": {
            "type": "object",
            "properties": {
                "maxAge": {
                    "description": "Duration string like \"24h\"",
                    "type": "string"
                },
                "maxBytes": {
                    "type": "integer"
                },
                "maxConsumers": {
                    "type": "integer"
                },
                "maxMsgs": {
                    "type": "integer"
                },
                "noAck": {
                    "type": "boolean"
                },
                "replicas": {
                    "type": "integer"
                },
                "retention": {
                    "description": "\"limits\", \"interest\", or \"workqueue\"",
                    "type": "string"
                },
                "subjects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and your API key.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}